refactor tower into apps
